Refactor Contact component imports to use Contact.interfaces for improved code organization; update Jest configuration to exclude type and interface files from coverage collection; enhance mock data imports for consistency.
